requirements.txt
fastapi
uvicorn
streamlit
requests
langgraph
openai
python-multipart
pydantic
whisper
ffmpeg-python
pytesseract
Pillow
opencv-python
pypdf
youtube-transcript-api
python-dotenv
google-cloud-vision
httpx